refactor(web): split WordBankScreen into thin orchestrator per ADR 0006
Same pattern as WordBankEntryDetail — the file contained 3 utility
functions (groupByDate, parseTags, groupByLetter), 2 sub-components
(SectionHeader, EntryGroup), and a constant (ALPHABET) alongside the
main screen component.

Extracted into:

  screens/WordBankScreen/
    index.tsx            ← thin orchestrator (~170 lines)
    EntryGroup.tsx       ← entry card link with tags
    SectionHeader.tsx    ← date/letter section divider
    utils.ts             ← groupByDate, parseTags, groupByLetter, ALPHABET

Import path unchanged — App.tsx resolves to index.tsx automatically.
